requirements.txt
##Notebook--Colab

ultralytics==8.0.140 ##install separately using pip install --user ultraytics to avoid installation error
numpy==1.19.5
pandas==0.25.3
seaborn==0.11.1
matplotlib==3.4.3
scikit-learn==1.0.2
pickleshare==0.7.5
tqdm==4.36.1
joblib==0.13.2
keras==2.5.0rc0
tensorflow==2.5.0
protobuf==3.20.0
h5py==2.10.0
Pillow==6.2.0
opencv-python==4.1.1.26
opencv-contrib-python==4.1.1.26
torch==1.13.1
torchvision==0.14.1


##Flask Framework
#gunicorn==20.1.0

Flask==1.1.1
Jinja2==2.10.3
Markdown==3.4.1
MarkupSafe==2.0.1
Werkzeug==0.16.0
itsdangerous==1.1.0
glob2==0.7
db-sqlite3==0.0.1



####If Jupyter NOtebook is not working then install this packages

#pip install --upgrade jupyterhub
#pip install --upgrade --user nbconvert